# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing passengers and taxists in a ride-sharing application.",
        "title": "Ride-Sharing Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.55.42:8080",
    "basePath": "/",
    "paths": {
        "/distances": {
            "get": {
                "description": "get all distances",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "GET distances",
                "responses": {}
            },
            "post": {
                "description": "Add a new distance between two places",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Add a new distance between two places",
                "parameters": [
                    {
                        "description": "Distance data",
                        "name": "distance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DistanceRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/places": {
            "get": {
                "description": "get all places",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "GET car places",
                "responses": {}
            },
            "post": {
                "description": "add a new place",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "POST a new place",
                "parameters": [
                    {
                        "description": "Announcement",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PlacesRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/protected/distances/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "DELETE distance by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "DELETE distance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Distance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/protected/places/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "DELETE place by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "DELETE place",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/protected/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Get user profile",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "handlers.PlacesRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Ashgabat"
                }
            }
        },
...
        "models.DistanceRequest": {
            "type": "object",
            "required": [
                "distance",
                "from_place",
                "to_place"
            ],
            "properties": {
                "distance": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 100
                },
                "from_place": {
                    "type": "integer",
                    "example": 1
                },
                "to_place": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}